pyproject.toml
[project]
name = "imnfs"
version = "0.1.0"
description = "Interactive model with NumPy, Pandas, and Streamlit"
authors = [
    { name = "AlexNhat", email = "anhnhat909940@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10,<3.12"

dependencies = [
    "numpy==1.26.4",
    "pandas==2.2.3",
    "openpyxl==3.1.5",
    "streamlit==1.39.0",
    "latexify-py (>=0.4.4,<0.5.0)",
    "matplotlib (>=3.10.7,<4.0.0)"
]

[project.optional-dependencies]
dev = [
    "ipykernel==6.29.5",
    "loguru==0.7.2",
    "rich==13.9.2"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
